+++ swlibs/texus2/lib/rgt.c
@@ -49,7 +49,7 @@ typedef struct _rgtHeader{
 
 
 
-static void swapShorts(unsigned short *array, long length)
+static void swapShorts(unsigned short *array, int length)
 {
     unsigned short s;
     while (length--) {
@@ -59,7 +59,7 @@ static void swapShorts(unsigned short *array, long length)
 }
 
 #if 0 /* not used */
-static void swapLongs(unsigned int *array, long length)
+static void swapLongs(unsigned int *array, int length)
 {
     unsigned int s;
     while (length--) {
@@ -71,7 +71,7 @@ static void swapLongs(unsigned int *array, long length)
 #endif
 
 // just swap RGB into BGR (leave MSB undefined)
-static void swapRGB(unsigned int *array, long length)
+static void swapRGB(unsigned int *array, int length)
 {
     unsigned int s;
     while (length--) {
@@ -163,7 +163,7 @@ _txReadRGTData( FILE *stream, TxMip *info)
 
 #if 1
         if (swap) {
-            swapRGB((unsigned int *)data32, (long)info->width);
+            swapRGB((unsigned int *)data32, (int)info->width);
         }
 #endif
